+++ packages/eas-cli/src/__tests__/commands/build-version-sync-test.ts
@@ -1,5 +1,4 @@
 import { Workflow } from '@expo/eas-build-job';
-import { AppVersionSource, EasJson } from '@expo/eas-json';
 import fs from 'fs-extra';
 import path from 'path';
 
@@ -9,6 +8,8 @@ import {
   mockCommandContext,
   mockProjectId,
   mockTestCommand,
+  withLocalVersionSource,
+  withRemoteVersionSource,
 } from './utils';
 import { updateNativeVersionsAsync as updateAndroidNativeVersionsAsync } from '../../build/android/version';
 import { updateNativeVersionsAsync as updateIosNativeVersionsAsync } from '../../build/ios/version';
@@ -19,6 +20,7 @@ import { AppQuery } from '../../graphql/queries/AppQuery';
 import { AppVersionQuery } from '../../graphql/queries/AppVersionQuery';
 import { getAppBuildGradleAsync } from '../../project/android/gradleUtils';
 import { resolveTargetsAsync } from '../../project/ios/target';
+import { AppVersionSourceUpdateOption } from '../../project/remoteVersionSource';
 import { resolveWorkflowAsync } from '../../project/workflow';
 import * as prompts from '../../prompts';
 
@@ -37,16 +39,6 @@ jest.mock('../../log');
 jest.mock('../../prompts');
 jest.mock('../../utils/json');
 
-function withRemoteVersionSource(easJson: EasJson): EasJson {
-  return {
-    ...easJson,
-    cli: {
-      ...easJson.cli,
-      appVersionSource: AppVersionSource.REMOTE,
-    },
-  };
-}
-
 describe(BuildVersionSyncView, () => {
   afterEach(() => {
     jest.clearAllMocks();
@@ -64,6 +56,34 @@ describe(BuildVersionSyncView, () => {
     jest.mocked(prompts.promptAsync).mockImplementation(async () => ({
       version: '1000',
     }));
+    jest.mocked(resolveWorkflowAsync).mockImplementation(async () => Workflow.MANAGED);
+
+    const cmd = mockTestCommand(BuildVersionSyncView, ['--platform=android'], ctx);
+    const syncAndroidAsync = jest.spyOn(cmd, 'syncAndroidAsync' as any);
+
+    await cmd.run();
+    expect(AppVersionQuery.latestVersionAsync).toHaveBeenCalledWith(
+      ctx.loggedIn.graphqlClient,
+      mockProjectId,
+      'ANDROID',
+      'eas.test.com'
+    );
+    expect(syncAndroidAsync).not.toHaveBeenCalled();
+  });
+
+  test('syncing version for managed project on platform android when appVersionSource is not set and the user chooses to set it to REMOTE', async () => {
+    const ctx = mockCommandContext(BuildVersionSyncView, {});
+    jest.mocked(AppQuery.byIdAsync).mockImplementation(async () => getMockAppFragment());
+    jest.mocked(AppVersionQuery.latestVersionAsync).mockImplementation(async () => ({
+      buildVersion: '1000',
+      storeVersion: '0.0.1',
+    }));
+    jest.mocked(prompts.promptAsync).mockImplementationOnce(async () => ({
+      version: '1000',
+    }));
+    jest
+      .mocked(prompts.selectAsync)
+      .mockImplementationOnce(async () => AppVersionSourceUpdateOption.SET_TO_REMOTE);
     jest.mocked(resolveWorkflowAsync).mockImplementation(async () => Workflow.MANAGED);
 
     const cmd = mockTestCommand(BuildVersionSyncView, ['--platform=android'], ctx);
@@ -179,7 +199,9 @@ describe(BuildVersionSyncView, () => {
   });
 
   test('syncing version aborts when appVersionSource is set to local and users refuse auto configuration', async () => {
-    const ctx = mockCommandContext(BuildVersionSyncView, {});
+    const ctx = mockCommandContext(BuildVersionSyncView, {
+      easJson: withLocalVersionSource(getMockEasJson()),
+    });
     jest.mocked(AppVersionQuery.latestVersionAsync).mockImplementation(async () => null);
     jest.mocked(AppQuery.byIdAsync).mockImplementation(async () => getMockAppFragment());
     jest.mocked(prompts.confirmAsync).mockImplementation(async () => false);
@@ -189,8 +211,24 @@ describe(BuildVersionSyncView, () => {
     expect(AppVersionMutation.createAppVersionAsync).not.toHaveBeenCalledWith();
   });
 
-  test('syncing version when appVersionSource is set to local and user allows auto configuration', async () => {
+  test('syncing version aborts when appVersionSource is not set and the user chooses to set it to LOCAL, and they refuse auto configuration', async () => {
     const ctx = mockCommandContext(BuildVersionSyncView, {});
+    jest.mocked(AppVersionQuery.latestVersionAsync).mockImplementation(async () => null);
+    jest.mocked(AppQuery.byIdAsync).mockImplementation(async () => getMockAppFragment());
+    jest
+      .mocked(prompts.selectAsync)
+      .mockImplementationOnce(async () => AppVersionSourceUpdateOption.SET_TO_LOCAL);
+    jest.mocked(prompts.confirmAsync).mockImplementation(async () => false);
+
+    const cmd = mockTestCommand(BuildVersionSyncView, ['--platform=android'], ctx);
+    await expect(cmd.run()).rejects.toThrowError('Aborting...');
+    expect(AppVersionMutation.createAppVersionAsync).not.toHaveBeenCalledWith();
+  });
+
+  test('syncing version when appVersionSource is set to local and user allows auto configuration', async () => {
+    const ctx = mockCommandContext(BuildVersionSyncView, {
+      easJson: withLocalVersionSource(getMockEasJson()),
+    });
     jest.mocked(AppVersionQuery.latestVersionAsync).mockImplementation(async () => null);
     jest.mocked(AppQuery.byIdAsync).mockImplementation(async () => getMockAppFragment());
     jest.mocked(prompts.confirmAsync).mockImplementation(async () => true);
@@ -200,5 +238,34 @@ describe(BuildVersionSyncView, () => {
 
     const easJsonAfterCmd = await fs.readJson(path.join(ctx.projectDir, 'eas.json'));
     expect(easJsonAfterCmd.cli.appVersionSource).toBe('remote');
+  });
+
+  test('syncing version when appVersionSource is not set and the user chooses to set it to LOCAL and they allow auto configuration', async () => {
+    const ctx = mockCommandContext(BuildVersionSyncView, {});
+    jest.mocked(AppVersionQuery.latestVersionAsync).mockImplementation(async () => null);
+    jest.mocked(AppQuery.byIdAsync).mockImplementation(async () => getMockAppFragment());
+    jest
+      .mocked(prompts.selectAsync)
+      .mockImplementationOnce(async () => AppVersionSourceUpdateOption.SET_TO_LOCAL);
+    jest.mocked(prompts.confirmAsync).mockImplementation(async () => true);
+
+    const cmd = mockTestCommand(BuildVersionSyncView, ['--platform=android'], ctx);
+    await cmd.run();
+
+    const easJsonAfterCmd = await fs.readJson(path.join(ctx.projectDir, 'eas.json'));
+    expect(easJsonAfterCmd.cli.appVersionSource).toBe('remote');
+  });
+
+  test('syncing version aborts when appVersionSource is not set and the user chooses to configure manually', async () => {
+    const ctx = mockCommandContext(BuildVersionSyncView, {});
+    jest.mocked(AppVersionQuery.latestVersionAsync).mockImplementation(async () => null);
+    jest.mocked(AppQuery.byIdAsync).mockImplementation(async () => getMockAppFragment());
+    jest
+      .mocked(prompts.selectAsync)
+      .mockImplementationOnce(async () => AppVersionSourceUpdateOption.ABORT);
+
+    const cmd = mockTestCommand(BuildVersionSyncView, ['--platform=android'], ctx);
+    await expect(cmd.run()).rejects.toThrowError('Aborted.');
+    expect(AppVersionMutation.createAppVersionAsync).not.toHaveBeenCalledWith();
   });
 });